+++ src/main/java/vanillaautomated/blockentities/BreakerBlockEntity.java
@@ -42,51 +42,12 @@ import java.util.Random;
 
 public class BreakerBlockEntity extends MachineBlockEntity implements SidedInventory, PropertyDelegateHolder {
 
-    private final PropertyDelegate propertyDelegate;
     public int speed = VanillaAutomated.config.breakerTime;
-    DefaultedList<ItemStack> items = DefaultedList.ofSize(2, ItemStack.EMPTY);
-    private int processingTime;
-    private int fuelTime;
-    private int maxFuelTime;
     private Random random = new Random();
 
     public BreakerBlockEntity(BlockPos pos, BlockState state) {
         super(VanillaAutomatedBlocks.breakerBlockEntity, pos, state);
-        this.propertyDelegate = new PropertyDelegate() {
-            public int get(int index) {
-                switch (index) {
-                    case 0:
-                        return fuelTime;
-                    case 1:
-                        return processingTime;
-                    case 2:
-                        return maxFuelTime;
-                    default:
-                        return 0;
-                }
-            }
-
-            public void set(int index, int value) {
-                switch (index) {
-                    case 0:
-                        fuelTime = value;
-                        break;
-                    case 1:
-                        processingTime = value;
-                        break;
-                    case 2:
-                        maxFuelTime = value;
-                        break;
-                    default:
-                        break;
-                }
-
-            }
-
-            public int size() {
-                return 4;
-            }
-        };
+        items = DefaultedList.ofSize(2, ItemStack.EMPTY);
     }
 
     public static boolean canUseAsFuel(ItemStack stack) {
@@ -151,30 +112,6 @@ public class BreakerBlockEntity extends MachineBlockEntity implements SidedInven
         }
     }
 
-    @Override
-    public void readNbt(NbtCompound tag) {
-        super.readNbt(tag);
-        Inventories.readNbt(tag, items);
-        if (tag.contains("CustomName", 8)) {
-            this.customName = Text.Serializer.fromJson(tag.getString("CustomName"));
-        }
-        this.processingTime = tag.getShort("ProcessingTime");
-        this.fuelTime = tag.getShort("FuelTime");
-        this.maxFuelTime = tag.getShort("MaxFuelTime");
-    }
-
-    @Override
-    public NbtCompound writeNbt(NbtCompound tag) {
-        Inventories.writeNbt(tag, items);
-        if (this.customName != null) {
-            tag.putString("CustomName", Text.Serializer.toJson(this.customName));
-        }
-        tag.putShort("ProcessingTime", (short) this.processingTime);
-        tag.putShort("FuelTime", (short) this.fuelTime);
-        tag.putShort("MaxFuelTime", (short) this.maxFuelTime);
-        return super.writeNbt(tag);
-    }
-
     public static void tick(World world, BlockPos blockPos, BlockState blockState, BreakerBlockEntity t) {
         if (world.isClient) {
             return;
@@ -196,7 +133,7 @@ public class BreakerBlockEntity extends MachineBlockEntity implements SidedInven
         }
 
         // Freeze when powered
-        if (world.getBlockState(t.getPos()).get(Properties.POWERED).booleanValue()) {
+        if (world.getBlockState(t.getPos()).get(Properties.POWERED)) {
             return;
         }
 
@@ -238,7 +175,7 @@ public class BreakerBlockEntity extends MachineBlockEntity implements SidedInven
     }
 
     private void generateItems() {
-        PlayerEntity player = world.getClosestPlayer((int) pos.getX(), (int) pos.getY(), (int) pos.getZ(), Float.MAX_VALUE, false);
+        PlayerEntity player = world.getClosestPlayer(pos.getX(), pos.getY(), pos.getZ(), Float.MAX_VALUE, false);
         if (player == null) {
             return;
         }
@@ -278,7 +215,7 @@ public class BreakerBlockEntity extends MachineBlockEntity implements SidedInven
             return 0;
         } else {
             Item item = fuel.getItem();
-            return (Integer) AbstractFurnaceBlockEntity.createFuelTimeMap().getOrDefault(item, 0);
+            return AbstractFurnaceBlockEntity.createFuelTimeMap().getOrDefault(item, 0);
         }
     }
