+++ src/main/java/vanillaautomated/blockentities/MobFarmBlockEntity.java
@@ -39,52 +39,13 @@ import java.util.Random;
 
 public class MobFarmBlockEntity extends MachineBlockEntity implements SidedInventory, PropertyDelegateHolder, Nameable {
 
-    private final PropertyDelegate propertyDelegate;
     public int speed = VanillaAutomated.config.mobFarmTime;
-    DefaultedList<ItemStack> items = DefaultedList.ofSize(11, ItemStack.EMPTY);
-    private int processingTime;
-    private int fuelTime;
-    private int maxFuelTime;
     private Random random = new Random();
     private String entityType = "";
 
     public MobFarmBlockEntity(BlockPos pos, BlockState state) {
         super(VanillaAutomatedBlocks.mobFarmBlockEntity, pos, state);
-        this.propertyDelegate = new PropertyDelegate() {
-            public int get(int index) {
-                switch (index) {
-                    case 0:
-                        return fuelTime;
-                    case 1:
-                        return processingTime;
-                    case 2:
-                        return maxFuelTime;
-                    default:
-                        return 0;
-                }
-            }
-
-            public void set(int index, int value) {
-                switch (index) {
-                    case 0:
-                        fuelTime = value;
-                        break;
-                    case 1:
-                        processingTime = value;
-                        break;
-                    case 2:
-                        maxFuelTime = value;
-                        break;
-                    default:
-                        break;
-                }
-
-            }
-
-            public int size() {
-                return 4;
-            }
-        };
+        items = DefaultedList.ofSize(11, ItemStack.EMPTY);
     }
 
     public static boolean canUseAsFuel(ItemStack stack) {
@@ -173,26 +134,12 @@ public class MobFarmBlockEntity extends MachineBlockEntity implements SidedInven
     @Override
     public void readNbt(NbtCompound tag) {
         super.readNbt(tag);
-        Inventories.readNbt(tag, items);
-        if (tag.contains("CustomName", 8)) {
-            this.customName = Text.Serializer.fromJson(tag.getString("CustomName"));
-        }
-        this.processingTime = tag.getShort("ProcessingTime");
-        this.fuelTime = tag.getShort("FuelTime");
-        this.maxFuelTime = tag.getShort("MaxFuelTime");
         this.entityType = tag.getString("EntityType");
         updateEntityType();
     }
 
     @Override
     public NbtCompound writeNbt(NbtCompound tag) {
-        Inventories.writeNbt(tag, items);
-        if (this.customName != null) {
-            tag.putString("CustomName", Text.Serializer.toJson(this.customName));
-        }
-        tag.putShort("ProcessingTime", (short) this.processingTime);
-        tag.putShort("FuelTime", (short) this.fuelTime);
-        tag.putShort("MaxFuelTime", (short) this.maxFuelTime);
         tag.putString("EntityType", this.entityType);
         return super.writeNbt(tag);
     }
@@ -208,7 +155,7 @@ public class MobFarmBlockEntity extends MachineBlockEntity implements SidedInven
         }
 
         // Freeze when powered
-        if (world.getBlockState(t.getPos()).get(Properties.POWERED).booleanValue()) {
+        if (world.getBlockState(t.getPos()).get(Properties.POWERED)) {
             if (t.isBurning()) {
                 t.fuelTime--;
             }
