+++ src/main/java/vanillaautomated/blockentities/PlacerBlockEntity.java
@@ -33,51 +33,12 @@ import java.util.Random;
 
 public class PlacerBlockEntity extends MachineBlockEntity implements SidedInventory, PropertyDelegateHolder, Nameable {
 
-    private final PropertyDelegate propertyDelegate;
     public int speed = VanillaAutomated.config.placerTime;
-    DefaultedList<ItemStack> items = DefaultedList.ofSize(2, ItemStack.EMPTY);
-    private int processingTime;
-    private int fuelTime;
-    private int maxFuelTime;
     private Random random = new Random();
 
     public PlacerBlockEntity(BlockPos pos, BlockState state) {
         super(VanillaAutomatedBlocks.placerBlockEntity, pos, state);
-        this.propertyDelegate = new PropertyDelegate() {
-            public int get(int index) {
-                switch (index) {
-                    case 0:
-                        return fuelTime;
-                    case 1:
-                        return processingTime;
-                    case 2:
-                        return maxFuelTime;
-                    default:
-                        return 0;
-                }
-            }
-
-            public void set(int index, int value) {
-                switch (index) {
-                    case 0:
-                        fuelTime = value;
-                        break;
-                    case 1:
-                        processingTime = value;
-                        break;
-                    case 2:
-                        maxFuelTime = value;
-                        break;
-                    default:
-                        break;
-                }
-
-            }
-
-            public int size() {
-                return 4;
-            }
-        };
+        items = DefaultedList.ofSize(2, ItemStack.EMPTY);
     }
 
     public static boolean canUseAsFuel(ItemStack stack) {
@@ -140,30 +101,6 @@ public class PlacerBlockEntity extends MachineBlockEntity implements SidedInvent
             ItemStack itemStack = this.items.get(1);
             return canUseAsFuel(stack) || stack.getItem() == Items.BUCKET && itemStack.getItem() != Items.BUCKET;
         }
-    }
-
-    @Override
-    public void readNbt(NbtCompound tag) {
-        super.readNbt(tag);
-        Inventories.readNbt(tag, items);
-        if (tag.contains("CustomName", 8)) {
-            this.customName = Text.Serializer.fromJson(tag.getString("CustomName"));
-        }
-        this.processingTime = tag.getShort("ProcessingTime");
-        this.fuelTime = tag.getShort("FuelTime");
-        this.maxFuelTime = tag.getShort("MaxFuelTime");
-    }
-
-    @Override
-    public NbtCompound writeNbt(NbtCompound tag) {
-        Inventories.writeNbt(tag, items);
-        if (this.customName != null) {
-            tag.putString("CustomName", Text.Serializer.toJson(this.customName));
-        }
-        tag.putShort("ProcessingTime", (short) this.processingTime);
-        tag.putShort("FuelTime", (short) this.fuelTime);
-        tag.putShort("MaxFuelTime", (short) this.maxFuelTime);
-        return super.writeNbt(tag);
     }
 
     public static void tick(World world, BlockPos blockPos, BlockState blockState, PlacerBlockEntity t) {
